% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_MMU_Block.R
\name{convert_MMU_Block}
\alias{convert_MMU_Block}
\title{Convert Measurement, methods and units block}
\usage{
convert_MMU_Block(methodColStrs = NULL, unitName = NULL, org.df,
  unit.df = NULL, verbose = FALSE)
}
\arguments{
\item{methodColStrs}{an array of strings idenfiying the methods}

\item{verbose}{}
}
\description{
This function will convert a long format with columns declared as methods, measurements, or units.
}
