% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertKeyedData.R
\name{convertKeyedData}
\alias{convertKeyedData}
\title{Convert data tables}
\usage{
convertKeyedData(data.ls, key.df, returnFormat = c("3Table", "long",
  "wide")[1])
}
\arguments{
\item{data.ls}{list containing the data tables specified in the key}

\item{key.df}{the key or look-up table that converts one formate to another}

\item{returnFormat}{string identifying the return type, can be either "long", "wide", or "3Table"}
}
\value{
a data frame of one of three formats: a long table, wide table, or 
3 data tables specifying the sample, field, and study data
}
\description{
This function converts a set of data tables from a specified format to a keyed format 
including renaming variables.
}
\examples{
\dontrun{
rawData <- readCanandaUplandForest2018() #~7Mb
longData <- convertKeyedData(rawData, rawData$ISCN5_key, returnFormat = 'long') #~50Mb
wideData <- convertKeyedData(rawData, rawData$ISCN5_key, returnFormat = 'wide') #~7Mb
3TableData <- convertKeyedData(rawData, rawData$ISCN5_key, returnFormat = '3Table') #~9Mb
}
}
