% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCN3.R
\name{ISCN3}
\alias{ISCN3}
\title{Load ISCN Layer and Meta data}
\usage{
ISCN3(dataDir = NULL, orginalFormat = FALSE, verbose = FALSE)
}
\arguments{
\item{dataDir}{path to the folder containing ISCN_ALL_DATA_LAYER_C*_1-1.xlsx, ISCN_ALL-DATA-CITATION_1-1.xlsx and ISCN_ALL_DATA_DATASET_1-1.xlsx files. If this is left NULL then files will be downloaded to a temporary directory from the ISCN website and then deleted.}

\item{verbose}{boolean flag denoting whether or not to print lots of status messages}
}
\value{
list of data.table with layer, dataset and citation information
}
\description{
This function first downloads the layer, profile, citation, and dataset tables from the ISCN website (http://iscn.fluxdata.org/data/access-data/database-reports/) data available: ftp://ftp.fluxdata.org/.deba/ISCN/ALL-DATA/* It then either returns the original structure or reformats the data.
}
