% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCN3_3.R
\name{ISCN3_3}
\alias{ISCN3_3}
\title{International Soil Carbon Network vs 3.3}
\usage{
ISCN3_3(data_dir, datasets_exclude = c(), verbose = FALSE)
}
\arguments{
\item{data_dir}{character refers to a directory where the data is located or downloaded to.}

\item{datasets_exclude}{A list of dataset names to exclude}

\item{verbose}{logical flag to print out messages or not.}
}
\value{
a list with the study table, layer table and profile table.
}
\description{
This function calls the ISCN3 function to pull the ISCN 3_2 archive off of EDI. 
It cleans these tables by recasting columns to appropriate data types, and then replacing ISCN computations with NA values. 
It then returns the reformatted data.
}
