% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCN4.R
\name{ISCN4}
\alias{ISCN4}
\title{Read in two data sets queued in for data processing for ISCN4.}
\usage{
ISCN4(
  dataDir = NULL,
  orginalFormat = FALSE,
  onlyNewData = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dataDir}{string identifing the data directory to download data to, or NULL to download to a temporary directory}

\item{onlyNewData}{boolean flagging new ISCN4 data only or appended to ISCN3}

\item{verbose}{boolean flagging verbose error messages}
}
\value{
a list of dataframes containing the study, field, and sample data
}
\description{
Read in two data sets queued in for data processing for ISCN4.
}
\examples{
\dontrun{
temp <- ISCN4()
}
}
