% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCN5.R
\name{ISCN5}
\alias{ISCN5}
\title{Read in new ISCN5 contributions}
\usage{
ISCN5(dataDir, orginalFormat = TRUE, newDataOnly = TRUE,
  verbose = FALSE)
}
\arguments{
\item{dataDir}{string identifying the data directory}

\item{orginalFormat}{boolean flag whether or not to translate to keep in orginal data table.}
}
\value{
a list of data tables
}
\description{
Read in the two ISCN5 contributions
}
