% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ISCN_key_list}
\alias{ISCN_key_list}
\title{List of data model dictionaries}
\format{
list of data tables:
\describe{
  \item{afsda}{adfa}
  \item{adfas}{fdasasd}
  ...
}
}
\usage{
ISCN_key_list
}
\description{
A list contianing data model dicitionaries.
}
\keyword{datasets}
