% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeKeys.R
\name{makeKeys}
\alias{makeKeys}
\title{Translational keys for datasets}
\usage{
makeKeys(dataDir = NULL, saveDir = NULL)
}
\value{
a list of data tables
}
\description{
Translational keys for datasets
}
\examples{
allKeys <- makeKeys()
}
