% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputData_PowellCenter.R
\name{outputData_PowellCenter}
\alias{outputData_PowellCenter}
\title{Write out data to the Powell Center Working Group format as specified by a template file.}
\usage{
outputData_PowellCenter(data.df, outputDir, templatefile)
}
\arguments{
\item{data.df}{data frame with internal formating and a column (ISCN_id) specifying the harnomized measure names}

\item{outputDir}{target directory string for the outputs}

\item{templatefile}{string pointing to a template file}
}
\value{
returns a list of dataframes that reflect the output
}
\description{
WARNING currently there is no QA/QC to enforce unit harmonziation.
}
