% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData_emptyStandard.R
\name{processData_emptyStandard}
\alias{processData_emptyStandard}
\title{Empty data skeleton}
\usage{
processData_emptyStandard()
}
\value{
A list that reflects the structure of the data representation
}
\description{
This function returns an empty list of data frames with the internal standard fields.
You can either use this as a basis to build new ingestion scripts and this is the
 structure that scripts should be tested against in testthat checks.
 This is the minimum structure.
}
\examples{
emptyFramework <- processData_emptyStandard()
}
