% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData_ISCN3.R
\name{processData_ISCN3}
\alias{processData_ISCN3}
\title{Load ISCN vs 3 Layer data}
\usage{
processData_ISCN3(dir = "repoData/ISCN_3", verbose = FALSE)
}
\arguments{
\item{dir}{string identifying the folder the data is in. Should have a subfolder 'Layers' containing all ISCN_ALL_DATA_LAYER_C[1-4]_1-1.csv files.}

\item{verbose}{boolean flag for verbose outputs}
}
\value{
a data frame with all data in the correct format
}
\description{
ISCN vs 3 (http://iscn.fluxdata.org/data/access-data/database-reports/) data available: ftp://ftp.fluxdata.org/.deba/ISCN/ALL-DATA/ISCN_ALL_DATA_LAYER_C1_1-1.xlsx ftp://ftp.fluxdata.org/.deba/ISCN/ALL-DATA/ISCN_ALL_DATA_LAYER_C2_1-1.xlsx ftp://ftp.fluxdata.org/.deba/ISCN/ALL-DATA/ISCN_ALL_DATA_LAYER_C3_1-1.xlsx ftp://ftp.fluxdata.org/.deba/ISCN/ALL-DATA/ISCN_ALL_DATA_LAYER_C4_1-1.xlsx
}
