% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData_ISCN3.R
\name{processMethodBlock_ISCN3}
\alias{processMethodBlock_ISCN3}
\title{Process a methods block}
\usage{
processMethodBlock_ISCN3(methodNames = NULL, unitName = NULL, sampleTemp,
  unit.df = NULL, verbose = FALSE)
}
\arguments{
\item{methodNames}{strings identifying the metods name for the block}

\item{unitName}{strings identifying the unit name for the block}

\item{sampleTemp}{stirngs identifying the samples for the block}

\item{unit.df}{data frame for the unit for the block keyed on the measure}

\item{verbose}{boolean if you want lots of error messages}
}
\value{
a list of data frames with the sample data and measurement information
}
\description{
Force the samples to discard repeats
}
\details{
TODO Update this to use convert_MMU_Block
}
