% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData_ISCN3.R
\name{processWorksheet_ISCN3}
\alias{processWorksheet_ISCN3}
\title{Process each ISCN worksheet}
\usage{
processWorksheet_ISCN3(csvFile = "Layers/ISCN_ALL_DATA_LAYER_C1_1-1.csv",
  verbose = TRUE)
}
\arguments{
\item{csvFile}{string identifying the CSV file}

\item{verbose}{boolean if you want LOTS of messages}
}
\value{
a list of data frames
}
\description{
Process each ISCN worksheet
}
