% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBuell2011.R
\name{readBuell2011}
\alias{readBuell2011}
\title{Raw read for USGS_S3C (2011)}
\usage{
readBuell2011(data_dir, download = TRUE, verbose = FALSE)
}
\arguments{
\item{data_dir}{filename for download directory}

\item{download}{boolean that will download the files from repository}
}
\value{
a data frame containing newly ingested data from the Myers-Smith dataset
}
\description{
Download and read in raw data from Myers-Smith, Isla. 2005. Soil data for cores from a transect from the center of the BBC collapse scar into the surrounding burn, Bonanza Creek LTER - University of Alaska Fairbanks. BNZ:192, http://www.lter.uaf.edu/data/data-detail/id/192. doi:10.6073/pasta/b0e9120983438c27bf1a30f37f0e5200
}
\details{
This is not post-processed or QA/QCed by this package.
}
