% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCanadaUplandForest2018.R
\name{readCanandaUplandForest2018}
\alias{readCanandaUplandForest2018}
\title{Raw read for Canadian upland soils (2018)}
\usage{
readCanandaUplandForest2018(dataDir, download = TRUE)
}
\arguments{
\item{dataDir}{filename for download directory}

\item{download}{boolean that will download the files from repository}
}
\value{
a list with meta data and raw table reads. This will be 4.6 Mb in size.
}
\description{
Download and read in the raw data tables from Shaw, C., Hilger, A., Filiatrault, M. and
 Kurz, W. (2018), A Canadian upland forest soil profile and carbon stocks 
 database. Ecology. doi:10.1002/ecy.2159
 This is not post-processed or QA/QCed by this package.
}
