% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCPEAT.R
\name{readCPEAT}
\alias{readCPEAT}
\title{CPEAT project reads}
\usage{
readCPEAT(dataDir)
}
\arguments{
\item{dataDir}{identify the download directory}
}
\value{
a list of data frames, the first data frame with the meta data and
 a second data frame with the records
}
\description{
This reads in the specified records of the CPEAT project. Currently under development, not all
the metadata is parsed
}
