% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKane2004.R
\name{readKane2004}
\alias{readKane2004}
\title{Raw Read for data corresponding to the Kane dataset}
\usage{
readKane2004(dataDir, download = TRUE, verbose = FALSE)
}
\arguments{
\item{dataDir}{filename for download directory}

\item{download}{boolean that will download the files from repository}
}
\value{
a list with meta data and raw table reads. This will be 49 KB in size.
}
\description{
Download and read in the raw data tables from Kane, E., Ping, C., Chapin, S. and 
Valentine, D. with Bonanza LTER. doi:10.6073/pasta/d09433eee2cb6587eca672864cf7e90f
}
\details{
This is not post-processed or QA/QCed by this package.
}
