% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKeyedData.R
\name{readKeyedData}
\alias{readKeyedData}
\title{This function will read in a generic data file and convert it a long data table and a wide data table with the assistance of a key. This is intended for relatively small data sets, if your data is more then 10K layers (with an expected data load >10Mb) consider writing your own function.}
\usage{
readKeyedData(filename = NA, key.df = NA, verticalTable = c("metadata"),
  excludeSheets = c("controlled vocabulary"), skipRows = 0, dropRows = NA,
  verbose = FALSE)
}
\arguments{
\item{filename}{characters identifying an Excel file with worksheets or set of csv files to be read in}

\item{key.df}{data frame converting headers to variables.}

\item{verticalTable}{a list of sheets or file names with data by column instead of by row. Typically this is metadata where a single entry is expected.}

\item{excludeSheets}{list of sheets in Excel workbook not to read in}

\item{skipRows}{the rows (or columns) to skip because they are human readable descriptions or units. The header will be read as being after these rows.}

\item{dropRows}{rows to be dropped after the data is read in. Typically these are descritions below the headers.}

\item{verbose}{a flag to print out helpful statements}
}
\value{
a list of a wide and long data tables with a copy of the final key
}
\description{
This function will read in a generic data file and convert it a long data table and a wide data table with the assistance of a key. This is intended for relatively small data sets, if your data is more then 10K layers (with an expected data load >10Mb) consider writing your own function.
}
