% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMyersSmith2005.R
\name{readMyersSmith2005}
\alias{readMyersSmith2005}
\title{Raw read for Myers-Smith (2005)}
\usage{
readMyersSmith2005(dataDir, download = TRUE, verbose = FALSE)
}
\arguments{
\item{dataDir}{filename for download directory}

\item{download}{boolean that will download the files from repository}
}
\value{
a list with meta data and raw table reads. This will be 67 KB in size.
}
\description{
Download and read in raw data from Myers-Smith, Isla. 2005. Soil data for cores from a transect from the center of the BBC collapse scar into the surrounding burn, Bonanza Creek LTER - University of Alaska Fairbanks. BNZ:192, http://www.lter.uaf.edu/data/data-detail/id/192. doi:10.6073/pasta/b0e9120983438c27bf1a30f37f0e5200
}
\details{
This is not post-processed or QA/QCed by this package.
}
