% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSchuur2009.R
\name{readSchuur2009}
\alias{readSchuur2009}
\title{Raw Read for data corresponding to the Schuur dataset}
\usage{
readSchuur2009(dataDir, download = TRUE, verbose = FALSE)
}
\arguments{
\item{dataDir}{filename for download directory}

\item{download}{boolean that will download the files from repository}
}
\value{
a list with meta data and raw table reads. This will be ____ KB in size.
}
\description{
Download and read in the raw data tables from Hicks-Pries C. and Bonanza Creek LTER (2009).
Identifier: knb-lter-bnz.366.16
}
\details{
This is not post-processed or QA/QCed by this package.
}
