% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUSGSHarden2008.R
\name{readUSGSHarden2008}
\alias{readUSGSHarden2008}
\title{Raw Read for data corresponding to the USGS_Harden dataset}
\usage{
readUSGSHarden2008(dataDir, download = TRUE, verbose = FALSE)
}
\arguments{
\item{dataDir}{filename for download directory}

\item{download}{boolean that will download the files from repository}
}
\value{
a list with meta data and raw table reads. This will be ____ KB in size.
}
\description{
Download and read in the raw data tables from Harden, W., Turetsky, M., Manies, P., and
Waldrop, M. with Bonanza LTER. doi:10.6073/pasta/9b363cc74bcd784cfcee85e6920f2989
}
\details{
This is not post-processed or QA/QCed by this package.
}
