% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilDepthAvgToBound.R
\name{soilDepthAvgToBound}
\alias{soilDepthAvgToBound}
\title{Title of function here}
\usage{
soilDepthAvgToBound(soilDepthAvg, topRef = 0)
}
\arguments{
\item{soilDepthAvg}{}

\item{topRef}{}
}
\value{

}
\description{
description here This is a test function that converts soil average depths into soil depth bounds.
}
