% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData_ISCN3.R
\name{processData_ISCN3}
\alias{processData_ISCN3}
\title{Primary function to load all of ISCN3 data}
\usage{
processData_ISCN3(dir = "repoData/ISCN_3")
}
\arguments{
\item{dir}{string identifying the folder the data is in. Should have a subfolder 'Layers' containing all ISCN_ALL_DATA_LAYER_C*_1-1.csv files.}
}
\value{
a data frame with all data in the correct format
}
\description{
Primary function to load all of ISCN3 data
}
