% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readISRaD.R
\name{readISRaD}
\alias{readISRaD}
\title{Read ISRaD contribution down to the layer level.}
\usage{
readISRaD(dataDir, verbose = FALSE)
}
\arguments{
\item{dataDir}{string identifying the file where the ISRaD data will be downloaded}

\item{verbose}{boolean flag to produce excessive debugging statements}
}
\value{
a list of character-based data tables
}
\description{
Read ISRaD contribution down to the layer level.
}
